pyproject.toml
[project]
name = "trip-tale"
version = "0.1.0"
description = ""
authors = [
    {name = "tkmsikeda",email = "tkms.ikeda@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "piexif (>=1.1.3,<2.0.0)"
]

[tool.ruff]
select = ["E", "F", "N"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
